Consolidate test suite into cohesive structure
- Move comprehensive data type tests from tests/data_types.rs to appropriate unit test modules
- Add data type, conversion, and array tests to src/data.rs (9 new tests)
- Add new data type attribute tests to src/das.rs (2 new tests)
- Add new data type array and constructor tests to src/dds.rs (2 new tests)
- Remove redundant tests/data_types.rs file
- Maintain tests/parse.rs for integration tests with actual data files
- All 22 tests passing (19 unit + 3 integration) with cohesive organization
